docs(aio): update reactive-forms to CLI (#20019)
PR Close #20019
