+++ aio/content/examples/reactive-forms/src/app/app.component.html
@@ -1,0 +1,4 @@
+<div class="container">
+  <h1>Reactive Forms</h1>
+  <app-hero-list></app-hero-list>
+</div>
