+++ aio/content/examples/reactive-forms/src/app/hero-detail/hero-detail-6.component.ts
@@ -5,7 +5,7 @@ import { Component, Input, OnChanges }             from '@angular/core';
 import { FormBuilder, FormGroup, Validators }      from '@angular/forms';
 
 // #docregion import-hero
-import { Hero, states } from './data-model';
+import { Hero, states } from '../data-model';
 // #enddocregion import-hero
 
 ////////// 6 ////////////////////
