+++ aio/content/examples/reactive-forms/src/app/hero-detail/hero-detail-7.component.ts
@@ -5,7 +5,7 @@ import { Component, Input, OnChanges }             from '@angular/core';
 import { FormBuilder, FormGroup, Validators }      from '@angular/forms';
 
 // #docregion import-address
-import { Address, Hero, states } from './data-model';
+import { Address, Hero, states } from '../data-model';
 // #enddocregion import-address
 
 // #enddocregion imports
