+++ aio/content/examples/reactive-forms/src/index-final.html
@@ -5,10 +5,8 @@
     <title>Hero Form</title>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="styles.css">
-     <!-- #docregion bootstrap -->
     <link rel="stylesheet" href="https://unpkg.com/bootstrap@3.3.7/dist/css/bootstrap.min.css">
-     <!-- #enddocregion bootstrap -->
+    <link rel="stylesheet" href="styles.css">
 
      <script src="node_modules/core-js/client/shim.min.js"></script>
