+++ aio/content/examples/reactive-forms/src/index.html
@@ -5,9 +5,7 @@
     <title>Hero Form</title>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-     <!-- #docregion bootstrap -->
     <link rel="stylesheet" href="https://unpkg.com/bootstrap@3.3.7/dist/css/bootstrap.min.css">
-     <!-- #enddocregion bootstrap -->
   </head>
 
   <body>
